Remove minified versions of index.js
After some research, it seems that it would be better to leave index.js
un-minified. Bandwidth is not an issue when it comes to server-side
javascript (node.js), and people using this package on the web will
probably be using webpack, browserify, etc... Also, this library is so
small that adding an extra build step is probably not worth it.

Still though, thanks @MaXwellFalstein! You reminded me to minify the
json... :)