refactor(es2015): Add es2015 support with babel
Changes many files to use es2015. Setup tests to work with it a well.
Istanbul refuses to generate coverage, however, so that needs to be fixed.
All files are built into the dist/ folder. This folder is not commited to
github, as its in the .gitignore, but Travis CI will still publish it to
npm because it doesn't exist in the .npmignore.